fikset '-' i package
