+++ core/pom.xml
@@ -3,23 +3,19 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <artifactId>prosjekt</artifactId>
-    <groupId>org.gruppe3-mmfs</groupId>
+    <groupId>org.gruppe3.mmfs</groupId>
     <version>1.0-SNAPSHOT</version>
   </parent>
-
-  <groupId>org.gruppe3-mmfs</groupId>
+  <groupId>org.gruppe3.mmfs</groupId>
   <artifactId>core</artifactId>
   <version>1.0-SNAPSHOT</version>
-
   <name>core</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
-
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.release>17</maven.compiler.release>
   </properties>
-
   <dependencyManagement>
     <dependencies>
       <dependency>
@@ -31,7 +27,6 @@
       </dependency>
     </dependencies>
   </dependencyManagement>
-
   <dependencies>
     <dependency>
       <groupId>org.junit.jupiter</groupId>
@@ -45,9 +40,9 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
-
   <build>
-    <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
+    <pluginManagement>
+      <!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
       <plugins>
         <!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
         <plugin>
