+++ core/src/test/java/org/gruppe3/mmfs/AppTest.java
@@ -1,0 +1,15 @@
+package org.gruppe3.mmfs;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+/** Unit test for simple App. */
+public class AppTest {
+
+  /** Rigorous Test :-) */
+  @Test
+  public void shouldAnswerWithTrue() {
+    assertTrue(true);
+  }
+}
