--- storage/src/main/java/org/gruppe3-mmfs/App.java
@@ -1,10 +1,0 @@
-package org.gruppe3-mmfs;
-
-/**
- * Hello world!
- */
-public class App {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
