+++ storage/src/main/java/org/gruppe3/mmfs/App.java
@@ -1,0 +1,8 @@
+package org.gruppe3.mmfs;
+
+/** Hello world! */
+public class App {
+  public static void main(String[] args) {
+    System.out.println("Hello World!");
+  }
+}
